Team view

+++ src/app/pages/teams-list/teams-list.component.ts
@@ -9,6 +9,7 @@ import { User } from '../../models/User';
 import { TeamService } from 'src/app/team/team.service';
 import { UserService } from 'src/app/user.service';
 import { environment } from 'src/environments/environment';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -31,6 +32,8 @@ export class TeamsListComponent implements OnInit {
   members?: []  
   currentUser?: User;
   requestDone: boolean = false;
+  users: Array<User> = [];
+
 
   // list members
   userList = [];
@@ -40,7 +43,7 @@ export class TeamsListComponent implements OnInit {
     private router: Router,
     private fb: FormBuilder,
     private modalService: NgbModal,
-    private http: HttpClient,
+    private http: HttpClient, 
     private userService: UserService,
   ) { }
 
@@ -48,6 +51,7 @@ export class TeamsListComponent implements OnInit {
 
     try {
       await this.CurrentUser();
+      await this.onFetchGroups();
     }
     catch (e) {
       this.requestDone = true;
@@ -58,34 +62,47 @@ export class TeamsListComponent implements OnInit {
     // get list of users (members)
     this.getUsersList();
 
+    console.log("******")
+     console.log(this.defaultMemberId);
+
     this.selectForm = this.fb.group({
       members: []
     });
     this.createGroupForm = this.fb.group({
-      name: [''],
-      members: ['']
+      name: ['']
     });
 
   }
 
+  async onFetchGroups() {
+    this.teams = await this.teamService.getAllGroup().toPromise();
+    this.users = await this.userService.findAll().toPromise();
+    this.defaultMemberId = this.teams.map(team => this.defaultUser(team.members));
+    this.requestDone = true;
+  }
+
+  defaultUser(members: Array<User>) {
+    if (members.length) {
+      return (members[0]._id);
+    }
+    return ("");
+  }
 
   // get current User 
   async CurrentUser() {
     this.currentUser = await this.userService.CurrentUser();
-    console.log("enter");
-    console.log(this.currentUser);
   }
 
 
   // update group
-  async update(team: Team) {
-     const toUpdate: Team = {
-       _id: team._id,
-      members: team.members.map((g: User) => g._id)
-    };
+  // async update(team: Team) {
+  //    const toUpdate: Team = {
+  //      _id: team._id,
+  //     members: team.members.map((g: User) => g._id)
+  //   };
 
-    await this.teamService.update(toUpdate).toPromise();
-  }
+  //   await this.teamService.update(toUpdate).toPromise();
+  // }
 
   
 
@@ -105,7 +122,7 @@ export class TeamsListComponent implements OnInit {
   // Fetching users data 
   getUsersList() {
     this.http
-      .get<any>(environment.baseUrl+this.ADD_URL)
+      .get<any>(environment.baseUrl+this.ADD_URL, { headers: { 'Authorization': `${localStorage.getItem('token')}` } })
       .subscribe(response => {
         this.userList = response.map((o: { search_label: string; firstname: string; lastname: string; email: string; }) => {
           o.search_label =
@@ -113,8 +130,7 @@ export class TeamsListComponent implements OnInit {
           `
           return o
         });
-        console.log(this.userList);
-
+        console.log(this.userList)
       }, error => {
         console.log(error);
       });
@@ -127,8 +143,6 @@ export class TeamsListComponent implements OnInit {
       centered: true,
       backdrop: 'static'
      });
-     console.log("current user : "+this.currentUser?._id);
-
    }
 
    // function used to select members added to group
@@ -153,6 +167,7 @@ export class TeamsListComponent implements OnInit {
 
   // clear list selected 
   clearListSelected() {
+    console.log("selected clear")
     this.selectForm.get('members')?.patchValue([]);
   }
 
@@ -166,7 +181,6 @@ export class TeamsListComponent implements OnInit {
     } as Team;
     this.teamService.createGroup(newGroup).subscribe(
       async (response: any) => {
-        console.log(response);
         this.getAllTeams();
         this.router.navigate(['/teams']);
       },
@@ -176,17 +190,42 @@ export class TeamsListComponent implements OnInit {
   }
    
   // delete a group
-   onDelete(id: string): void {
-    this.teamService.deleteGroup(id)
-      .subscribe(
-        (response: any) => {
-          console.log(response);
+  onDelete(id: string, teamM: Team): void {
+    const lengthMembers = Object.keys(teamM).length;
+    if(lengthMembers){
+      Swal.fire('Can\'t delete', 'Group has a members!!!', 'error');
+    } else {
+      Swal.fire({
+        title: 'Are you sure?',
+        text: 'You will not be able to recover this group!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, keep it'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.teamService.deleteGroup(id)
+          .subscribe(
+           (response: any) => {
+          Swal.fire('successfully deleted!', 'The group  has been deleted.', 'success')
           this.getAllTeams();
           this.router.navigate(['/teams']);
         },
         (error: any) => {
           console.log(error);
         });
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          Swal.fire(
+            'Cancelled',
+            'Your group  isn\'t deleted :)',
+            'error'
+          )
+        }
+      })
+      
+      
+    }
+    
   }
